Update vpc-peering-security-groups.md
The use of the word "was" on line #75 is awkward.  I reworded the sentence so that "is stale" is used.  The sentence might be grammatically correct but I think it is awkward to use "was stale" and the phrase "and then create".  The reason being is that at the time a new vpc connection is created the rule "is stale" rather than "was stale". 

Thank you for taking the time to consider my proposed change.